% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{Meteorological data for Southeast Asia}
\format{A data frame with 3,319,182 rows and 16 variables:
\itemize{
 \item \code{station}: climatic station ID
 \item \code{day}: date of data colletion
 \item \code{ta}: average temperature (°C)
 \item \code{tx}: maximum temperature (°C)
 \item \code{tn}: minimum temperature (°C)
 \item \code{slp}: atmospheric pressure at sea level (hPa)
 \item \code{h}: average relative humidity (%)
 \item \code{pp}: total rainfall and / or snowmelt (mm)
 \item \code{vv}: average visibility (km)
 \item \code{v}: average wind speed (km / h)
 \item \code{vm}: maximum sustained wind speed (km / h)
 \item \code{vg}: maximum speed of wind (km / h)
 \item \code{ra}: boolean indicating whether there was rain or drizzle
 \item \code{sn}: boolean indicating whether it snowed
 \item \code{ts}: boolean indicating whether there were storm
 \item \code{fg}: boolean indicating whether there was fod
}}
\source{
\href{https://en.tutiempo.net}{en.tutiempo.net}.
}
\usage{
meteo
}
\description{
A data containing the daily meteorological data for 7 countries in Southeast
Asia, from  \href{https://en.tutiempo.net}{en.tutiempo.net}.
}
\details{
The variable \code{station} is a key shared with dataset
\code{\link{stations}}.
}
\keyword{datasets}

