% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Locations and elevations of climatic stations}
\format{A data frame with 261 rows and 7 variables:
\itemize{
 \item \code{station}: climatic station ID
 \item \code{country}: country of the climatic station
 \item \code{location}: location of the climatic station
 \item \code{longitude}: longitude of the climatic station (decimal coordinates)
 \item \code{latitude}: latitude of the climatic station (decimal coordinates)
 \item \code{elevation}: elevation of the climatic station (m)
 \item \code{from}: year from which the data started being collected
}}
\source{
\href{https://en.tutiempo.net}{en.tutiempo.net}.
}
\usage{
stations
}
\description{
A dataset containing the coordinates and elevations of the climatic stations
contained in the \code{meteo} data frame.
}
\details{
The variable \code{station} is a key shared with dataset
\code{\link{meteo}}.
}
\keyword{datasets}

