% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutiempo.R
\docType{package}
\name{tutiempo}
\alias{tutiempo}
\alias{tutiempo-package}
\title{tutiempo: Meteorological data for southeast Asia.}
\description{
The tutiempo package provides daily times series for 7 countries in southeast
Asia
}
